refactor(frontend): move WearEffect into effects/
WearEffect and CoffeeStain are both decorative overlays; group them
in effects/. Update Card's imports and retarget the stylelint
rgba-exemption to the new path.
