+++ frontend/src/lib/components/effects/WearEffect.svelte
